third_party/blink/web_tests/external/wpt/css/cssom-view/element-scroll-promises-expected.txt

This is a testharness.js-based test.
[FAIL] Element.scroll(auto): return type
  assert_true: expected true got false
[FAIL] Element.scroll(instant): return type
  assert_true: expected true got false
[FAIL] Element.scroll(smooth): return type
  assert_true: expected true got false
[FAIL] Element.scroll(smooth): promise behavior
  assert_equals: Position after await expected 5000 but got 0
[FAIL] Element.scrollTo(auto): return type
  assert_true: expected true got false
[FAIL] Element.scrollTo(instant): return type
  assert_true: expected true got false
[FAIL] Element.scrollTo(smooth): return type
  assert_true: expected true got false
[FAIL] Element.scrollTo(smooth): promise behavior
  assert_equals: Position after await expected 5000 but got 0
[FAIL] Element.scrollBy(auto): return type
  assert_true: expected true got false
[FAIL] Element.scrollBy(instant): return type
  assert_true: expected true got false
[FAIL] Element.scrollBy(smooth): return type
  assert_true: expected true got false
[FAIL] Element.scrollBy(smooth): promise behavior
  assert_equals: Position after await expected 5000 but got 0
[FAIL] Element.scrollIntoView(auto): return type
  assert_true: expected true got false
[FAIL] Element.scrollIntoView(instant): return type
  assert_true: expected true got false
[FAIL] Element.scrollIntoView(smooth): return type
  assert_true: expected true got false
[FAIL] Element.scrollIntoView(smooth): promise behavior
  assert_equals: Position after await expected 5000 but got 0
Harness: the test ran to completion.
